refactor(platform-wallet): move token actions onto IdentityWallet facade
Token *actions* (transfer/burn/mint/claim/freeze/unfreeze/destroy_frozen_funds/
pause/resume/set_price/purchase/update_config) are identity-as-actor
operations — every method takes an `Identifier` for the actor identity,
and signing pulls a key off that identity. Hosting them on `TokenWallet`
(a sibling facade to `IdentityWallet`) forced FFI callers to keep
juggling `(wallet_handle, identity_id)` separately and put token
actions out of step with every other identity operation.

Mirror the DashPay merge that already happened on this codebase
(wallet/identity/network/mod.rs:7-15): move the actions onto
`IdentityWallet`, keep the wallet-level token *bookkeeping*
(watch/unwatch/sync/balance) on `TokenWallet`.

Layout after this refactor:

    wallet/tokens/
      wallet.rs                 # bookkeeping ONLY
      group_queries.rs          # token-contract-scoped reads (unchanged)

    wallet/identity/network/tokens/   # NEW — 12 actions live here
      mod.rs
      helpers.rs                # fetch_data_contract, resolve_signing_key,
                                # resolve_identity_and_signer (moved from
                                # TokenWallet, now pub(super) on IdentityWallet)
      transfer.rs               # token_transfer{,_with_signer,_with_external_signer}
      burn.rs                   # token_burn...
      mint.rs
      claim.rs
      freeze.rs
      unfreeze.rs
      destroy_frozen_funds.rs
      pause.rs
      resume.rs
      set_price.rs
      purchase.rs
      update_config.rs

Method naming follows the established IdentityWallet convention seen in
transfer.rs / withdrawal.rs / dpns.rs:

    token_<action>_with_external_signer  (preferred for external signers)
    token_<action>_with_signer
    token_<action>                        (legacy helpers; constructs
                                          IdentitySigner from manager)

`token_` prefix disambiguates from the existing `transfer_credits` family
on IdentityWallet.

FFI shims (rs-platform-wallet-ffi/src/tokens/*.rs) re-pointed:

    let token_wallet = wallet.tokens().clone();
    token_wallet.<action>_external_signer(identity_id, ...)
    →
    let identity_wallet = wallet.identity().clone();
    identity_wallet.token_<action>_with_external_signer(identity_id, ...)

The 14 C-ABI symbols are byte-identical before vs after — Swift wrappers
need no changes, the SwiftExampleApp builds clean with -warnings-as-errors,
and 110 + 69 platform-wallet / platform-wallet-ffi tests pass.

The two FFI query files (group_queries.rs) keep calling wallet.tokens()
because pending_group_actions / group_action_signers are token-contract-
scoped reads, not identity-scoped.
