+++ packages/rs-platform-wallet-ffi/src/tokens/burn.rs
@@ -1,4 +1,4 @@
-//! FFI binding for `TokenWallet::burn_external_signer`.
+//! FFI binding for `IdentityWallet::token_burn_with_external_signer`.
 //!
 //! Burn supports group-gated execution; the caller passes a flat
 //! `(group_info_kind, position, action_id, action_is_proposer)` tuple
@@ -115,11 +115,11 @@ pub unsafe extern "C" fn platform_wallet_token_burn(
 
     PLATFORM_WALLET_STORAGE
         .with_item(wallet_handle, |wallet| {
-            let token_wallet = wallet.tokens().clone();
+            let identity_wallet = wallet.identity().clone();
             let result = block_on_worker(async move {
                 let signer: &VTableSigner = &*(signer_addr as *const VTableSigner);
-                token_wallet
-                    .burn_external_signer(
+                identity_wallet
+                    .token_burn_with_external_signer(
                         id,
                         contract_id,
                         token_position,
