+++ packages/rs-platform-wallet-ffi/src/tokens/claim.rs
@@ -1,4 +1,4 @@
-//! FFI binding for `TokenWallet::claim_external_signer`.
+//! FFI binding for `IdentityWallet::token_claim_with_external_signer`.
 //!
 //! Claim is not group-gated — there's no `group_info_*` payload here.
 //! It does take a distribution-type discriminant, surfaced as a `u8`.
@@ -120,11 +120,11 @@ pub unsafe extern "C" fn platform_wallet_token_claim(
 
     PLATFORM_WALLET_STORAGE
         .with_item(wallet_handle, |wallet| {
-            let token_wallet = wallet.tokens().clone();
+            let identity_wallet = wallet.identity().clone();
             let result = block_on_worker(async move {
                 let signer: &VTableSigner = &*(signer_addr as *const VTableSigner);
-                token_wallet
-                    .claim_external_signer(
+                identity_wallet
+                    .token_claim_with_external_signer(
                         id,
                         contract_id,
                         token_position,
