+++ packages/rs-platform-wallet-ffi/src/tokens/mint.rs
@@ -1,4 +1,4 @@
-//! FFI binding for `TokenWallet::mint_external_signer`.
+//! FFI binding for `IdentityWallet::token_mint_with_external_signer`.
 //!
 //! Mint supports group-gated execution; the caller passes a flat
 //! `(group_info_kind, position, action_id, action_is_proposer)` tuple
@@ -139,11 +139,11 @@ pub unsafe extern "C" fn platform_wallet_token_mint(
 
     PLATFORM_WALLET_STORAGE
         .with_item(wallet_handle, |wallet| {
-            let token_wallet = wallet.tokens().clone();
+            let identity_wallet = wallet.identity().clone();
             let result = block_on_worker(async move {
                 let signer: &VTableSigner = &*(signer_addr as *const VTableSigner);
-                token_wallet
-                    .mint_external_signer(
+                identity_wallet
+                    .token_mint_with_external_signer(
                         id,
                         contract_id,
                         token_position,
