+++ packages/rs-platform-wallet-ffi/src/tokens/set_price.rs
@@ -1,4 +1,4 @@
-//! FFI binding for `TokenWallet::set_price_external_signer`.
+//! FFI binding for `IdentityWallet::token_set_price_with_external_signer`.
 //!
 //! Set Price configures the direct-purchase pricing schedule for a
 //! token. This entry point exposes only the flat-price form: a
@@ -125,11 +125,11 @@ pub unsafe extern "C" fn platform_wallet_token_set_price(
 
     PLATFORM_WALLET_STORAGE
         .with_item(wallet_handle, |wallet| {
-            let token_wallet = wallet.tokens().clone();
+            let identity_wallet = wallet.identity().clone();
             let result = block_on_worker(async move {
                 let signer: &VTableSigner = &*(signer_addr as *const VTableSigner);
-                token_wallet
-                    .set_price_external_signer(
+                identity_wallet
+                    .token_set_price_with_external_signer(
                         id,
                         contract_id,
                         token_position,
