+++ packages/rs-platform-wallet-ffi/src/tokens/transfer.rs
@@ -1,4 +1,4 @@
-//! FFI binding for `TokenWallet::transfer_external_signer`.
+//! FFI binding for `IdentityWallet::token_transfer_with_external_signer`.
 //!
 //! Transfers are never group-gated, so there's no `group_info_*`
 //! payload here.
@@ -117,11 +117,11 @@ pub unsafe extern "C" fn platform_wallet_token_transfer(
 
     PLATFORM_WALLET_STORAGE
         .with_item(wallet_handle, |wallet| {
-            let token_wallet = wallet.tokens().clone();
+            let identity_wallet = wallet.identity().clone();
             let result = block_on_worker(async move {
                 let signer: &VTableSigner = &*(signer_addr as *const VTableSigner);
-                token_wallet
-                    .transfer_external_signer(
+                identity_wallet
+                    .token_transfer_with_external_signer(
                         from_id,
                         contract_id,
                         token_position,
